proyecto utilizando pewee migrations/requirements.txt
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
ciscoconfparse==1.5.30
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
future==0.18.2
idna==2.10
Jinja2==3.0.1
junos-eznc==2.6.1
lxml==4.6.3
MarkupSafe==2.0.1
napalm==3.3.1
ncclient==0.6.9
netaddr==0.8.0
netmiko==3.4.0
ntc-templates==2.1.0
paramiko==2.7.2
passlib==1.7.4
peewee==3.14.4
peewee-migrations==0.3.19
psycopg2-binary==2.9.1
pycparser==2.20
pyeapi==0.8.4
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.5
PyYAML==5.4.1
requests==2.25.1
scp==0.13.4
six==1.16.0
tenacity==7.0.0
textfsm==1.1.0
transitions==0.8.8
urllib3==1.26.6
yamlordereddictloader==0.4.0
